Add workarounds for $playing YouTube videos
Lavalink sucks and their YouTube implementation is getting more and more
broken as time goes on. I've made some workarounds for the time being.
It takes a long time to get the actual video if the workaround has to be
used, but there's nothing that can be done about that.